+++ pycroscope/test_typeddict.py
@@ -1,6 +1,6 @@
 # static analysis: ignore
 from .test_name_check_visitor import TestNameCheckVisitorBase
-from .test_node_visitor import assert_passes, skip_if_not_installed
+from .test_node_visitor import assert_passes, skip_before, skip_if_not_installed
 from .value import (
     AnySource,
     AnyValue,
@@ -360,6 +360,29 @@ class TestTypedDict(TestNameCheckVisitorBase):
             value: T
 
         def capybara(movie: Movie) -> str:
+            return movie["director"]["name"]
+
+    @skip_before((3, 11))
+    @assert_passes(run_in_both_module_modes=True)
+    def test_stdlib_typeddict_class_syntax_and_newtype_base(self):
+        from typing import Generic, NewType, TypedDict, TypeVar
+
+        class Movie(TypedDict):
+            director: "Person"
+
+        class Person(TypedDict):
+            name: str
+
+        T = TypeVar("T")
+
+        class GenericMovie(TypedDict, Generic[T]):
+            value: T
+
+        BadMovieId = NewType("BadMovieId", Movie)  # E: incompatible_call
+
+        def capybara(movie: Movie, generic_movie: GenericMovie[int]) -> str:
+            print(BadMovieId)
+            print(generic_movie)
             return movie["director"]["name"]
 
     @assert_passes()